+++ cdm_reader_mapper/mdf_reader/write.py
@@ -12,6 +12,21 @@ import pandas as pd
 from cdm_reader_mapper.common import get_filename
 
 
+def _update_dtypes(dtypes, columns):
+    if isinstance(dtypes, dict):
+        dtypes = {k: v for k, v in dtypes.items() if k in columns}
+    return dtypes
+
+
+def _update_col_names(dtypes, col_o, col_n):
+    if isinstance(dtypes, str):
+        return dtypes
+    if col_o in dtypes.keys():
+        dtypes[col_n] = dtypes[col_o]
+        del dtypes[col_o]
+    return dtypes
+
+
 def write_data(
     data,
     mask=None,
@@ -108,14 +123,12 @@ def write_data(
         mask = mask[col_subset]
 
     header = []
-    info["dtypes"] = {k: v for k, v in info["dtypes"].items() if k in data.columns}
+    info["dtypes"] = _update_dtypes(info["dtypes"], data_df.columns)
     for col in data.columns:
         col_ = _join(col)
         header.append(col_)
-
-        if col in info["dtypes"]:
-            info["dtypes"][col_] = info["dtypes"][col]
-            del info["dtypes"][col]
+        info["dtypes"] = _update_col_names(info["dtypes"], col, col_)
+   
     info["parse_dates"] = [
         parse_date for parse_date in info["parse_dates"] if parse_date in header
     ]
@@ -128,7 +141,9 @@ def write_data(
         "sep": delimiter,
     }
     data.to_csv(os.path.join(out_dir, filename_data), **kwargs)
-    mask.to_csv(os.path.join(out_dir, filename_mask), **kwargs)
+    if not mask.empty:
+        mask.to_csv(os.path.join(out_dir, filename_mask), **kwargs)
+
 
     if info:
         with open(os.path.join(out_dir, filename_info), "w") as fileObj: